requirements.txt
prettytable>=0.7.2
tabulate>=0.8.7
pyfiglet>=0.7
pyyaml>=6.0
